PQueue.Prio.Private: extracted data type and instance definitions from PQueue.Prio
This way, we do no longer need to expose the PQ constructor.
