+++ pqueue.cabal
@@ -37,8 +37,9 @@ Library {
     Data.PQueue.Min
     Data.PQueue.Max
   other-modules:
-    Data.PQueue.Prio.Internals
     Data.PQueue.Internals
+    Data.PQueue.Prio.Private
+    Data.PQueue.Prio.Internals
     Data.PQueue.Prio.Max.Internals
     Data.PQueue.Top
     Control.Applicative.Identity
